fix bug by get scanTasks is 500 when search by useCaseId

+++ src/main/java/com/thoughtworks/security/scpapi/controller/request/TaskSearchRequest.java
@@ -13,7 +13,7 @@ public class TaskSearchRequest {
     private Integer appId;
     private Integer toolId;
     private Integer projectId;
-    private Integer useCaseId;
+    private Long useCaseId;
 
     private ScanTaskEnum status;
 }
